Split utils module into several helpers modules
