+++ src/utils/utils.js
@@ -2,8 +2,7 @@ import L from 'leaflet';
 import marker2x from 'leaflet/dist/images/marker-icon-2x.png';
 import marker from 'leaflet/dist/images/marker-icon.png';
 import markerShadow from 'leaflet/dist/images/marker-shadow.png';
-import { max, min, range } from 'ramda';
-import { useState } from 'react';
+import { range } from 'ramda';
 
 const TEN_ROUNDING_NUMBER = 10;
 const DEFAULT_TIMEOUT_DELAY = 2000;
@@ -17,16 +16,6 @@ export const stateFlagTimeout = (setTimeout) => (
 ) => {
   setState({ [flagName]: initialValue });
   setTimeout(() => setState({ [flagName]: !initialValue }), delay);
-};
-
-export const useStateFlagTimeout = (setTimeout) => (initialValue = true, delay = DEFAULT_TIMEOUT_DELAY) => {
-  const [ flag, setFlag ] = useState(initialValue);
-  const callback = () => {
-    setFlag(!initialValue);
-    setTimeout(() => setFlag(initialValue), delay);
-  };
-
-  return [ flag, callback ];
 };
 
 export const determineOrderDir = (clickedField, currentOrderField, currentOrderDir) => {
@@ -56,34 +45,3 @@ export const rangeOf = (size, mappingFn, startAt = 1) => range(startAt, size + 1
 
 export const roundTen = (number) => ceil(number / TEN_ROUNDING_NUMBER) * TEN_ROUNDING_NUMBER;
 
-export const useToggle = (initialValue = false) => {
-  const [ flag, setFlag ] = useState(initialValue);
-
-  return [ flag, () => setFlag(!flag) ];
-};
-
-export const formatDate = (format = 'YYYY-MM-DD') => (date) => date && date.format ? date.format(format) : date;
-
-export const formatIsoDate = (date) => date && date.format ? date.format() : date;
-
-export const ELLIPSIS = '...';
-
-export const progressivePagination = (currentPage, pageCount) => {
-  const delta = 2;
-  const pages = range(
-    max(delta, currentPage - delta),
-    min(pageCount - 1, currentPage + delta) + 1,
-  );
-
-  if (currentPage - delta > delta) {
-    pages.unshift(ELLIPSIS);
-  }
-  if (currentPage + delta < pageCount - 1) {
-    pages.push(ELLIPSIS);
-  }
-
-  pages.unshift(1);
-  pages.push(pageCount);
-
-  return pages;
-};
